+++ packages/ionic-react/src/schematics/application/schema.d.ts
@@ -13,7 +13,6 @@ export interface ApplicationSchematicSchema {
   classComponent?: boolean;
   skipWorkspaceJson?: boolean;
   js?: boolean;
-  disableSanitizer: boolean;
   capacitor: boolean;
 }